requirement.txt
numpy==1.15.4
pandas==0.23.4
Pillow==5.3.0
torch==0.4.1
torchvision==0.2.1

